Reorganize plot::items file structure
Changes:
* Move implementation details of items under sub-modules in items
* Outsource helper structs to 'values' module
* Simplify stacking implementation, add f64::ord() impl
* Rename HoverElem -> ClosestElem
